remove playground leftovers
