+++ vue3_poc/src/pages/HomePage.vue
@@ -1,89 +1,29 @@
 <template>
-	<div>
-		<div>
-			<h2 class="text-h2">
-				{{ persistentValue }}
-				<v-icon :icon="mdiHome"></v-icon>
-				<v-icon :icon="bcAssignment"></v-icon>
-			</h2>
-			<h2>{{ reactiveValue }}</h2>
-			<br />
-			<div>
-				<v-text-field
-					v-model="reactiveValue"
-					variant="outlined"
-					label="value"
-				/>
-			</div>
-			<br />
-			<div>
-				<v-btn color="primary" @click="onIncreaseCounter">
-					Counter in the component {{ counter }}
-				</v-btn>
-			</div>
-			<br />
-			<div>
-				<v-btn @click="onIncreaseStoreCounter">
-					Counter in the store {{ storeCounter }}
-				</v-btn>
-			</div>
-			<br />
-			<div>
-				<v-btn @click="dialogIsOpen = true">Open Dialog</v-btn>
-			</div>
-		</div>
-	</div>
-	<v-custom-dialog
-		v-model="dialogIsOpen"
-		has-buttons
-		:buttons="['close']"
-		@dialog-closed="onCloseDialog"
-	>
-		<template #title><h4>Dialog Title</h4></template>
-	</v-custom-dialog>
+	<v-container>
+		<v-row class="text-center">
+			<v-col cols="12">
+				<v-img :src="bird" class="my-3" contain height="400" />
+			</v-col>
+
+			<v-col class="mb-4">
+				<h1 class="display-2 font-weight-bold mb-3">
+					<h1>Welcome to Vue 3 POC</h1>
+				</h1>
+			</v-col>
+		</v-row>
+	</v-container>
 </template>
 
 <script lang="ts">
-import { mdiHome } from "@mdi/js";
-import { defineComponent, ref } from "vue";
-import { useCounterStore } from "@/store/counter";
-import { computed } from "vue";
-import vCustomDialog from "@/components/vCustomDialog.vue";
-import { bcAssignment } from "@/styles/icons";
+import { defineComponent } from "vue";
+
+import bird from "../assets/bird.svg";
+
 export default defineComponent({
-	components: {
-		vCustomDialog,
-	},
-	setup() {
-		const store = useCounterStore();
-		const persistentValue = "Composition api page";
-		const reactiveValue: object = ref("Reactive value");
-		const counter = ref(0);
-		const storeCounter = computed(() => store.counter);
-		const onIncreaseCounter = (): void => {
-			counter.value++;
-		};
-
-		const onIncreaseStoreCounter = (): void => {
-			store.increment();
-		};
-
-		const dialogIsOpen = ref(false);
-		const onCloseDialog = () => {
-			dialogIsOpen.value = false;
-		};
-
+	name: "HomePage",
+	data() {
 		return {
-			persistentValue,
-			reactiveValue,
-			onIncreaseCounter,
-			counter,
-			storeCounter,
-			onIncreaseStoreCounter,
-			dialogIsOpen,
-			onCloseDialog,
-			mdiHome,
-			bcAssignment,
+			bird,
 		};
 	},
 });
