--- vue3_poc/src/styles/variables.scss
@@ -1,3 +1,0 @@
-$border-radius: 6px;
-
-$h2-size: 2.75rem;
