+++ vue3_poc/src/styles/vuetify.scss
@@ -1,1 +1,3 @@
-@use "vuetify/styles";
+@use "vuetify/styles" as vuetify with (
+	$body-font-family: "PT Sans"
+);
